% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic.R
\name{get_country}
\alias{get_country}
\title{Get country}
\usage{
get_country(url_root)
}
\arguments{
\item{url_root}{Url root where the API is located.}
}
\value{
A data.frame, with the list of all countries.
}
\description{
function which gets a list of all countries available in the AClimate platform.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
df = get_countries(url_root)
print(head(df))

}
