% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast_climate_previous}
\alias{get_forecast_climate_previous}
\title{Get previous climate forecast}
\usage{
get_forecast_climate_previous(url_root, forecast, stations)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{forecast}{Represents the id of the forecast from which the information is desired.}

\item{stations}{Array of strings with the ids of the weather stations.}
}
\value{
A List with 3 attributes (probabilities, performance, scenarios).
}
\description{
Through this endpoint you can obtain the information obtained through the forecast process that is desired by means of the Id, the seasonal and subseasonal probabilities and the climatic scenarios. This endpoint is used through the Http GET method.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
forecast = "657006544afb9646da8c6b78"
stations=c("58504f1a006cb93ed40eebe2","58504f1a006cb93ed40eebe3")
obj_fp = get_forecast_climate_previous(url_root, forecast, stations)
print(obj_fp)

}
