% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast_crop_exc}
\alias{get_forecast_crop_exc}
\title{Get forecast crop exceedance}
\usage{
get_forecast_crop_exc(url_root, stations)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{stations}{Array of strings with the ids of the weather stations.}
}
\value{
A data.frame, with the crop forecast result for the weather stations.
}
\description{
Through this endpoint you can obtain the information obtained through of all forecast in the crop model process, yield data. This endpoint is used through the Http GET method.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
stations=c("58504f1a006cb93ed40eebe2","58504f1a006cb93ed40eebe3")
df = get_forecast_crop_exc(url_root, stations)
print(head(df))

}
