% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast_crop_previous}
\alias{get_forecast_crop_previous}
\title{Get crop forecast previous}
\usage{
get_forecast_crop_previous(url_root, forecast, stations)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{forecast}{Represents the id of the forecast from which the information is desired.}

\item{stations}{Array of strings with the ids of the weather stations.}
}
\value{
A data.frame, with the crop forecast result for the weather stations.
}
\description{
Through this endpoint you can obtains the forecast information for a specific forecast indicated in the parameters. This endpoint is used through the Http GET method.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
forecast = "657006544afb9646da8c6b78"
stations=c("58504f1a006cb93ed40eebe2","58504f1a006cb93ed40eebe3")
df = get_forecast_crop_previous(url_root, forecast, stations)
print(head(df))

}
