% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast_information}
\alias{get_forecast_information}
\title{Get forecast id}
\usage{
get_forecast_information(url_root, year)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{year}{Integer that represents the year from which you want to obtain the data for each forecast}
}
\value{
A data.frame, with the list of all forecast
}
\description{
Access primary forecast data for each month within a desired year using the HTTP GET method. This endpoint provides essential information, including unique identifiers, dates, and confidence levels, allowing for the identification of key details in the forecast for the specified year.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
year= 2023
forecasts = get_forecast_information(url_root, year)
print(forecasts)

}
