% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{get_forecast_subseasonal}
\alias{get_forecast_subseasonal}
\title{Get subseasonal forecast}
\usage{
get_forecast_subseasonal(url_root, stations)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{stations}{Array of strings with the ids of the weather stations.}
}
\value{
A data.frame, with the list of all subseasonal forecast data
}
\description{
Retrieve information from the subseasonal forecast process, including probabilities and climatic scenarios, using the HTTP GET method. This endpoint provides valuable insights into the forecast data and is accessed to gather relevant information.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
stations = c("63a3744005732d2a14260392","63a374ce05732d2a14260499")
subseasonal_data = get_forecast_subseasonal(url_root, stations)
print(subseasonal_data)

}
