% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoserver.R
\name{get_geo_polygon_name}
\alias{get_geo_polygon_name}
\title{Get geoserver polygon stores}
\usage{
get_geo_polygon_name(url_root, workspace)
}
\arguments{
\item{url_root}{Url root where the Geoserver is located.}

\item{workspace}{Name of the workspace from which the polygon datastores are to be obtained.}
}
\value{
A dataframe with polygons stores information.
}
\description{
Retrieves and lists all the polygon stores of a specific workspace using the HTTP GET method.
}
\examples{
url_root = "https://geo.aclimate.org/geoserver/"
workspace = "administrative"
obj_f = get_geo_polygon_name(url_root, workspace)
print(obj_f)

}
