% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoserver.R
\name{get_geo_polygons}
\alias{get_geo_polygons}
\title{Get geoserver shapefiles}
\usage{
get_geo_polygons(url_root, workspace, polygon_name)
}
\arguments{
\item{url_root}{Url root where the Geoserver is located.}

\item{workspace}{Name of the workspace from which the shapefiles are to be obtained.}

\item{polygon_name}{Name of the store from which the shapefiles are to be obtained.}
}
\value{
A shapefile with the polygon information or Null if an error is encountered, it will also be printed to the console.
}
\description{
Obtains the desired shapefile from the GeoServer using the HTTP GET method, by using workspace, and polygon name
}
\examples{
url_root = "https://geo.aclimate.org/geoserver/"
workspace = "administrative"
polygon_name = "ao_adm1"
obj_f = get_geo_polygons(url_root, workspace, polygon_name)
print(obj_f)

}
