% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoserver.R
\name{get_geo_stores}
\alias{get_geo_stores}
\title{Get geoserver stores}
\usage{
get_geo_stores(url_root, workspace)
}
\arguments{
\item{url_root}{Url root where the Geoserver is located.}

\item{workspace}{Name of the workspace from which the datastores are to be obtained.}
}
\value{
A dataframe with stores information.
}
\description{
Retrieves and lists all the stores of a specific workspace using the HTTP GET method.
}
\examples{
url_root = "https://geo.aclimate.org/geoserver/"
workspace = "climate_indices_pe"
obj_f = get_geo_stores(url_root, workspace)
print(obj_f)

}
