% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic.R
\name{get_geographic}
\alias{get_geographic}
\title{Get weather stations}
\usage{
get_geographic(url_root, country_id)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{country_id}{Id of the country}
}
\value{
A data.frame, with the list of all weather stations.
}
\description{
Retrieve detailed information on states, municipalities, weather stations, and crop-related details for a selected country. This endpoint, using the HTTP GET method, supports localized decision-making in agriculture.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
country_id = "61e59d829d5d2486e18d2ea8"
df = get_geographic(url_root, country_id)
print(head(df))

}
