% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic.R
\name{get_geographic_country}
\alias{get_geographic_country}
\title{Get country}
\usage{
get_geographic_country(url_root)
}
\arguments{
\item{url_root}{Url root where the API is located.}
}
\value{
A data.frame, with the list of all countries.
}
\description{
Obtain a global list of registered countries in the AClimate system using the HTTP GET method. Ideal for a broad overview of countries covered by AClimate.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
df = get_countries(url_root)
print(head(df))

}
