% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic.R
\name{get_geographic_crop}
\alias{get_geographic_crop}
\title{Get weather stations with crop information}
\usage{
get_geographic_crop(url_root, country_id)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{country_id}{Id of the country}
}
\value{
A data.frame, with the list of all weather stations.
}
\description{
Access crop-grouped data for states, municipalities, and meteorological stations within a chosen country. The endpoint, utilizing the HTTP GET method, offers insights for optimized decision-making based on specific crops.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
country = "61e59d829d5d2486e18d2ea8"
df = get_ws(url_root,country)
print(head(df))

}
