% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic.R
\name{get_geographic_ws}
\alias{get_geographic_ws}
\title{Get weather stations}
\usage{
get_geographic_ws(url_root, country_id)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{country_id}{Id of the country}
}
\value{
A list, with the list of all weather stations.
}
\description{
Retrieve detailed information weather stations, and crop-related details for a selected country. This endpoint, using the HTTP GET method, supports localized decision-making in agriculture.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
country = "61e59d829d5d2486e18d2ea8"
ws_list = get_geographic_ws(url_root,country)
print(head(ws_list))

}
