% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{get_historical_historicalclimatic}
\alias{get_historical_historicalclimatic}
\title{Get historical climate information}
\usage{
get_historical_historicalclimatic(url_root, stations)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{stations}{Array of strings with the ids of the weather stations.}
}
\value{
A data.frame, with the climatology of different climatic variables
}
\description{
Retrieve the weather history of a selected weather station using the HTTP GET method. This endpoint offers access to valuable historical climatic data.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
stations=c("58504f1a006cb93ed40eebe2","58504f1a006cb93ed40eebe3")
df = get_historical_historicalclimatic(url_root, stations)
print(df)

}
