% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical.R
\name{get_historical_historicalyieldyears}
\alias{get_historical_historicalyieldyears}
\title{Get years with historical crop performance data}
\usage{
get_historical_historicalyieldyears(url_root, stations)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{stations}{Array of strings with the ids of the weather stations.}
}
\value{
A data.frame, with the climatology of different climatic variables
}
\description{
Through this endpoint you can obtain the years that contain historical crop performance data of the selected weather station.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
stations=c("58504f1a006cb93ed40eebe2","58504f1a006cb93ed40eebe3")
df = get_historical_historicalyieldyears(url_root, stations)
print(df)

}
