% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic.R
\name{get_ws}
\alias{get_ws}
\title{Get weather stations}
\usage{
get_ws(url_root, country_id)
}
\arguments{
\item{url_root}{Url root where the API is located.}

\item{country_id}{Id of the country}
}
\value{
A data.frame, with the list of all weather stations.
}
\description{
function which gets a list of all weather stations available in the AClimate platform.
}
\examples{
url_root = "https://webapi.aclimate.org/api/"
country = "61e59d829d5d2486e18d2ea8"
df = get_ws(url_root,country)
print(head(df))

}
